requirements.txt
matplotlib==3.5.2
numpy==1.22.4
Pillow==9.1.1
streamlit==1.10.0
streamlit_drawable_canvas==0.9.1
tensorflow==2.9.1
#tensorflow_gpu==2.9.1
tqdm==4.64.0
